WEB-308 Split /careers-team-and-contact/ pagesection into two new pagesections:   /careers-and-team/, and /contact/. This seems much better.
