+++ numenta.com/config.toml
@@ -58,7 +58,8 @@ description = 'Leading the New Era of Machine Intelligence'
     business = '/business-strategy-and-ip/'
     anomaly = '/anomaly-detection-benchmark/'
     resources = '/papers-videos-and-more/'
-    careers = '/careers-team-and-contact/'
+    careers = '/careers-and-team/'
+    contact = '/contact/'
     # other site pages
     bami = '/biological-and-machine-intelligence/'
     blog = '/blog/'
