+++ numenta.com/package.json
@@ -1,6 +1,6 @@
 {
   "name": "Numenta.com",
-  "version": "0.2.89",
+  "version": "0.2.90",
   "description": "Numenta.com company website content, source code, and static generator tooling.",
   "license": "SEE LICENSE IN LICENSE.TXT",
   "main": false,
@@ -56,7 +56,7 @@
     "html2plaintext": "1.0.1",
     "js-yaml": "3.6.1",
     "lunr": "0.7.1",
-    "moment": "2.14.1",
+    "moment": "2.15.0",
     "ncp": "2.0.0",
     "platform": "1.3.1",
     "react": "15.3.1",
@@ -68,7 +68,7 @@
     "react-highlighter": "0.3.3",
     "react-icons": "2.2.1",
     "react-modal": "1.4.0",
-    "react-router": "2.7.0",
+    "react-router": "2.8.0",
     "react-tap-event-plugin": "1.0.0",
     "react-youtube": "7.1.1",
     "sitemap": "1.8.1",
