+++ numenta.com/pages/404.md
@@ -4,7 +4,7 @@ title: "Page Not Found"
 
 ### Sorry, the page you were looking for does not exist (404).
 
-* Please [Contact Us](/careers-team-and-contact/#contact) if you are having
-  problems, and we will respond as soon as possible.
+* Please [Contact Us](/contact/) if you are having problems, and we will respond
+  as soon as possible.
 
 * Visit the [Numenta.com homepage](/) for more information.
