+++ numenta.com/pages/500.md
@@ -4,7 +4,7 @@ title: "Website Server Error"
 
 ### Sorry, this website is experiencing technical difficulties (500).
 
-* Please [Contact Us](/careers-team-and-contact/#contact) if you are having
-  problems, and we will respond as soon as possible.
+* Please [Contact Us](/contact/) if you are having problems, and we will respond
+  as quickly as possible.
 
 * Visit the [Numenta.com homepage](/) for more information.
