+++ numenta.com/pages/careers-and-team/images/team/ed-colligan.jpg
