+++ numenta.com/pages/careers-and-team/images/team/ed-colligan@2x.jpg
