+++ numenta.com/pages/careers-and-team/images/team/harry-saal.jpg
