+++ numenta.com/pages/careers-and-team/images/team/jeff-hawkins.jpg
