+++ numenta.com/pages/careers-and-team/images/team/mike-farmwald.jpg
