+++ numenta.com/pages/careers-and-team/index.css
@@ -1,0 +1,13 @@
+.columns {
+  composes: layout-columns from '../../static/assets/css/theme.css';
+}
+
+.content {
+  composes: layout-content from '../../static/assets/css/theme.css';
+}
+
+.aside {
+  composes: mt2 from 'tachyons';
+  composes: tc from 'tachyons';
+  composes: layout-aside from '../../static/assets/css/theme.css';
+}
