+++ numenta.com/pages/careers-and-team/index.jsx
@@ -3,13 +3,13 @@ import React from 'react'
 
 import {scrollToSection} from '../../utils/client'
 
-import MainSections from '../_MainSections'
+import MainSections, {getNextSection} from '../_MainSections'
 import NextSection from '../../components/NextSection'
 import SectionCareers from './_Section'
 import Section from '../../components/Section'
 
 const Default = (<SectionCareers key="sectionCareers" />)
-const title = 'Careers, Team & Contact'
+const title = 'Careers & Team'
 
 
 /**
@@ -17,13 +17,18 @@ const title = 'Careers, Team & Contact'
  */
 class CareersPage extends React.Component {
 
-  state = {
-    sections: (
-      <Section headline={true} open={true} title={title}>
-        {Default}
-        <NextSection title="Back to Home" url="/" />
-      </Section>
-    ),
+  constructor(props) {
+    super(props)
+    const next = getNextSection(Default)
+
+    this.state = {
+      sections: (
+        <Section headline={true} open={true} title={title}>
+          {Default}
+          <NextSection {...next} />
+        </Section>
+      ),
+    }
   }
 
   componentDidMount() {
