+++ numenta.com/pages/contact/index.jsx
@@ -1,0 +1,54 @@
+import Helmet from 'react-helmet'
+import React from 'react'
+
+import {scrollToSection} from '../../utils/client'
+
+import MainSections from '../_MainSections'
+import NextSection from '../../components/NextSection'
+import SectionContact from './_Section'
+import Section from '../../components/Section'
+
+const Default = (<SectionContact key="sectionContact" />)
+const title = 'Contact'
+
+
+/**
+ *
+ */
+class ContactPage extends React.Component {
+
+  state = {
+    sections: (
+      <Section headline={true} open={true} title={title}>
+        {Default}
+        <NextSection title="Back to Home" url="/" />
+      </Section>
+    ),
+  }
+
+  componentDidMount() {
+    this.setState({  // eslint-disable-line react/no-did-mount-set-state
+      sections: (<MainSections current={Default} />),
+    })
+  }
+
+  componentDidUpdate() {
+    scrollToSection(global.document.getElementById(Default.key))
+  }
+
+  componentWillUnmount() {
+    this.setState({sections: []})
+  }
+
+  render() {
+    const {sections} = this.state
+    return (
+      <div>
+        <Helmet title={title} />
+        {sections}
+      </div>
+    )
+  }
+}
+
+export default ContactPage
