+++ numenta.com/pages/htm-studio/faq/whats-next.md
@@ -4,7 +4,7 @@ sort: c
 title: I am interested in integrating HTM technology into my application, what can I do next?
 ---
 
-[Contact Us](/careers-team-and-contact/#contact) to discuss adding HTM
-technology to your system and for licensing opportunities. You can also engage
-with our HTM open source community at http://numenta.org. There you will find
-all of our code, and learning resources to develop a project using HTM.
+[Contact Us](/contact/) to discuss adding HTM technology to your system and for
+licensing opportunities. You can also engage with our HTM open source community
+at http://numenta.org. There you will find all of our code, and learning
+resources to develop a project using HTM.
